Create core.magics.execution according to new API.
